# Conflicts:
#	i21Database/dbo/Stored Procedures/uspAPCreateVoucherDetailNonInvContract.sql
#	i21Database/dbo/Stored Procedures/uspCTReportBalance.sql
#	i21Database/dbo/Stored Procedures/uspMFReportItemPlan.sql
#	i21Database/dbo/Stored Procedures/uspMFReportItemPlanSummary.sql
#	i21Database/dbo/Tables/tblSMCompanyPreference.sql
#	i21Database/dbo/User Defined Types/VoucherDetailNonInvContract.sql
#	i21Database/i21Database.sqlproj

+++ i21Database/dbo/Stored Procedures/uspCTValidateContractHeader.sql
@@ -120,7 +120,7 @@ BEGIN TRY
 			intNoOfLoad					INT,
 			intFutureMarketId			INT,
 			intFutureMonthId			INT,
-			dblFutures					INT
+			dblFutures					NUMERIC(18,6)
 			
 	);
